+++ src/chess.zig
@@ -6,6 +6,7 @@ pub const Color = color.Color;
 pub const File = square.File;
 pub const Piece = @import("chess/piece.zig").Piece;
 pub const PieceList = @import("chess/PieceList.zig");
+pub const PieceMap = @import("chess/PieceMap.zig");
 pub const Placement = @import("chess/Placement.zig");
 pub const Rank = square.Rank;
 pub const RankFile = square.RankFile;