+++ src/tests.zig
@@ -10,6 +10,9 @@ comptime {
     _ = @import("chess/piece/ConditionImpl.zig");
     _ = @import("chess/piece/LookupImpl.zig");
     _ = @import("chess/PieceList.zig");
+    _ = @import("chess/PieceMap.zig");
+    _ = @import("chess/piece_map/LookupImpl.zig");
+    _ = @import("chess/piece_map/DirectImpl.zig");
     _ = @import("chess/Placement.zig");
     _ = @import("chess/square.zig");
 }