+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+#### Chess module
+
+- `PieceMap` board representation.
+
+
 ## [0.2.0] - 2024-08-12
 
 ### Added